+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerModule.kt
@@ -1,0 +1,7 @@
+package com.bitmovin.analytics.exoplayer
+
+enum class ExoPlayerModule(val moduleName: String) {
+    HLS("goog.exo.hls"),
+    DASH("goog.exo.dash"),
+    SMOOTH_STREAMING("goog.exo.smoothstreaming")
+}