+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Fixed
+- exoplayer collector crashes when specific exoplayer modules have not been loaded (AN-966)
+
 ## v1.11.0
 
 ### Added